fix: address runtime diagnostics review feedback
